* Modify DialogBoxRenderer by moving most logic from Base to Basic
