+++ include/AllegroFlare/Elements/DialogBoxRenderer.hpp
@@ -26,9 +26,7 @@ namespace AllegroFlare
          void render();
          void draw_special_state_empty_text(float width=1, float height=1);
          void draw_action_text(std::string text="[unset-action-text]", float width=1, float height=1);
-         void draw_styled_revealed_text(float max_width=1);
-         std::string obtain_dialog_box_text();
-         int obtain_dialog_box_num_revealed_characters();
+         void draw_styled_revealed_text(float max_width=1, std::string text="[unset-text]", int num_revealed_characters=999);
          std::string concat_text(std::string source_text="", int length=0);
          ALLEGRO_FONT* obtain_dialog_font();
       };
