+++ src/AllegroFlare/Elements/DialogBoxRenderer.cpp
@@ -7,15 +7,12 @@
 #include <allegro5/allegro_primitives.h>
 #include <allegro5/allegro_color.h>
 #include <AllegroFlare/Elements/DialogBoxFrame.hpp>
+#include <AllegroFlare/Elements/DialogBoxes/Basic.hpp>
 #include <AllegroFlare/Elements/DialogBoxes/Choice.hpp>
 #include <AllegroFlare/Elements/DialogBoxes/YouGotAnItem.hpp>
 #include <AllegroFlare/Elements/DialogBoxes/TitleText.hpp>
 #include <AllegroFlare/Elements/DialogBoxRenderers/ChoiceRenderer.hpp>
 #include <AllegroFlare/Elements/DialogBoxRenderers/YouGotAnItemRenderer.hpp>
-#include <stdexcept>
-#include <sstream>
-#include <stdexcept>
-#include <sstream>
 #include <stdexcept>
 #include <sstream>
 #include <stdexcept>
@@ -94,7 +91,8 @@ void DialogBoxRenderer::render()
       float width = place.size.x;
       float height = place.size.y;
       place.start_transform();
-      AllegroFlare::Elements::DialogBoxes::Choice* choice_dialog_box = static_cast<AllegroFlare::Elements::DialogBoxes::Choice*>(dialog_box);
+      AllegroFlare::Elements::DialogBoxes::Choice* choice_dialog_box =
+         static_cast<AllegroFlare::Elements::DialogBoxes::Choice*>(dialog_box);
       AllegroFlare::Elements::DialogBoxRenderers::ChoiceRenderer(font_bin, bitmap_bin, choice_dialog_box, width, height).render();
       place.restore_transform();
    }
@@ -136,17 +134,23 @@ void DialogBoxRenderer::render()
       place.start_transform();
       AllegroFlare::Elements::DialogBoxFrame(place.size.x, place.size.y).render();
 
-      if (dialog_box->get_finished())
+      AllegroFlare::Elements::DialogBoxes::Basic* basic_dialog_box =
+         dynamic_cast<AllegroFlare::Elements::DialogBoxes::Basic*>(dialog_box);
+
+      if (basic_dialog_box->get_finished())
       {
          draw_special_state_empty_text(place.size.x, place.size.y);
       }
       else
       {
-         draw_styled_revealed_text(place.size.x);
+         std::string dialog_current_page_text = basic_dialog_box->get_current_page_text();
+         int num_revealed_characters = basic_dialog_box->get_num_revealed_characters();
+
+         draw_styled_revealed_text(place.size.x, dialog_current_page_text, num_revealed_characters);
 
          // draw the player's action cursor thing at the bottom
-         int current_dialog_box_page_character_count = dialog_box->get_current_page_text().length();
-         int num_revealed_characters = obtain_dialog_box_num_revealed_characters();
+         int current_dialog_box_page_character_count = dialog_current_page_text.length();
+         //int num_revealed_characters = obtain_dialog_box_num_revealed_characters();
          if (num_revealed_characters >= current_dialog_box_page_character_count)
          {
             //if (dialog_box->at_last_page()) draw_action_text("[close]", place.size.x, place.size.y);
@@ -199,9 +203,8 @@ void DialogBoxRenderer::draw_action_text(std::string text, float width, float he
    return;
 }
 
-void DialogBoxRenderer::draw_styled_revealed_text(float max_width)
+void DialogBoxRenderer::draw_styled_revealed_text(float max_width, std::string text, int num_revealed_characters)
 {
-   std::string text = obtain_dialog_box_text();
    float text_padding_x = 40.0f;
    float text_padding_y = 30.0f;
    float text_box_max_width = max_width - (text_padding_x * 2);
@@ -209,7 +212,7 @@ void DialogBoxRenderer::draw_styled_revealed_text(float max_width)
    float line_height = al_get_font_line_height(text_font);
    //ALLEGRO_COLOR text_color = al_color_html("66a9bc");
    ALLEGRO_COLOR text_color = al_color_name("skyblue");
-   int num_revealed_characters = obtain_dialog_box_num_revealed_characters();
+   //int num_revealed_characters = obtain_dialog_box_num_revealed_characters();
 
    al_draw_multiline_text(
       text_font,
@@ -222,28 +225,6 @@ void DialogBoxRenderer::draw_styled_revealed_text(float max_width)
       concat_text(text, num_revealed_characters).c_str()
    );
    return;
-}
-
-std::string DialogBoxRenderer::obtain_dialog_box_text()
-{
-   if (!(dialog_box))
-      {
-         std::stringstream error_message;
-         error_message << "DialogBoxRenderer" << "::" << "obtain_dialog_box_text" << ": error: " << "guard \"dialog_box\" not met";
-         throw std::runtime_error(error_message.str());
-      }
-   return dialog_box->get_current_page_text();
-}
-
-int DialogBoxRenderer::obtain_dialog_box_num_revealed_characters()
-{
-   if (!(dialog_box))
-      {
-         std::stringstream error_message;
-         error_message << "DialogBoxRenderer" << "::" << "obtain_dialog_box_num_revealed_characters" << ": error: " << "guard \"dialog_box\" not met";
-         throw std::runtime_error(error_message.str());
-      }
-   return dialog_box->get_num_revealed_characters();
 }
 
 std::string DialogBoxRenderer::concat_text(std::string source_text, int length)
